+++ src/main/java/com/udemy/projeto/spring/services/impl/CategoriaServiceImpl.java
@@ -17,7 +17,7 @@ public class CategoriaServiceImpl implements CategoriaService {
     @Autowired
     private CategoriaRepository repository;
 
-    public Categoria buscar(long id) {
+    public Categoria buscar(Integer id) {
         Optional<Categoria> obj = repository.findById(id);
         return obj.orElseThrow(() -> new ObjectNotFoundException(
                 "Objeto não encontrado! Id: " + id + ", Tipo: " + Categoria.class.getName()));
